who v0.0.1 Who.ProPublicaAPI.Chamber

This module defines methods for querying the ProPublica API based on chamber of Congress (i.e. House or Senate)

Functions

get_members_by_chamber(chamber)

get_members_by_chamber(String.t) :: map

Get a list of current members in either the House or Senate

EX:

get_members_by_chamber("house")
get_members_by_chamber("senate")

get_recent_chamber_actions(chamber)

get_recent_chamber_actions(String.t) :: map

Get a list of recent floor actions by chamber

EX:

get_recent_chamber_actions("house")
get_recent_chamber_actions("senate")