who v0.0.1 Who.ProPublicaAPI.Committee

This module defines methods for querying the ProPublica API for information about specific Congressional committees and/or subcommittees

Functions

get_commitee_info(chamber)

get_commitee_info(String.t) :: list

Get a list of committees and subcommitees by Chamber.

EX: Who.ProPublicaAPI.Committee.get_commitee_info(“house”)

get_commitee_info(chamber, committee_id)

get_commitee_info(String.t, String.t) :: list

Query the API for information about a specific Committee in a given Chamber of Congress.

EX: Who.ProPublicaAPI.Committee.get_commitee_info(“house”, “hspw”)

get_committee_hearings(chamber, committee_id)

Get information about recent hearings held by a given Committee.

EX: Who.ProPublicaAPI.Committee.get_committee_hearings(“house”, “hspw”)