who v0.0.1 Who.ProPublicaAPI.Member

This module defines methods for querying the ProPublica API for individual member (i.e. Representative or Sentaor) information

Search options include:

Get Member by bio ID

Get member by state/chamber/district

Functions

get_member_by_id(member_id)

get_member_by_id(String.t) :: map

Query for Member by member_id as assigned by the Biographical Directory of the United States Congress.

http://bioguide.congress.gov/biosearch/biosearch.asp

EX: get_member_by_id(“K000388”)

get_member_by_state(chamber, state)

get_member_by_state(String.t, String.t) :: map

Query for Members from a specific state by chamber and/or district.

EX:

get_member_by_state("senate", "NJ")
get_member_by_state("house", "NJ")
get_member_by_state("house", "NJ", 12)

get_member_by_state(binary, state, district)

get_member_by_state(String.t, String.t, non_neg_integer) :: map

get_member_by_state(String.t, any, any) :: map

get_member_votes(member_id)

get_member_votes(String.t) :: map

Query for a Member’s voting positions by bio ID.

EX:

get_member_votes("K000388")