who v0.0.1 WhoWeb.BillPresenter

Takes in and processes data having to do with a given Bill and returns an object containing the aforementioned info.

WhoWeb.BillPresenter.new()

Types

t()

t() :: %WhoWeb.BillPresenter{bill_id: String.t, committees: [String.t], date_introduced: String.t, has_cosponsors: boolean, history: [map], latest_action: String.t, latest_action_date: String.t, num_of_cosponsors: integer, number: String.t, sponsor_id: String.t, sponsor_name: String.t, sponsor_party: String.t, sponsor_title: String.t, summary: String.t, title: String.t, url: String.t}

Functions

format_bill_id(bill_number)

Takes in a Bill number and returns a reformatted copy to use with ProPubica API requests.

EX:

format_bill_id(("H.R.21")
>>> "hr21"

new(bill)

new(map) :: nil | t

Creates a new %BillPresenter{} named struct.