who v0.0.1 WhoWeb.MemberView

Functions

__phoenix_recompile__?()

Returns true whenever the list of templates changes in the filesystem.

__resource__()

The resource name, as an atom, for this view

__templates__()

Returns the template root alongside all templates.

build_member_party(member)

build_member_party(String.t) :: String.t

Takes the “R” or “D” party flag from the Member JSON blob and returns the full party name as a string.

EX:

WhoWeb.MemberView.build_member_party("K000383")
>>> "Independent"

WhoWeb.MemberView.build_member_party("C001035")
>>> "Republican"

build_member_title(chamber)

build_member_title(String.t) :: String.t

Based on the Member’s chamber, this function returns the Member’s title

EX:

WhoWeb.MemberView.build_member_title("House")
>>> "Representative"

WhoWeb.MemberView.build_member_title("Senate")
>>> "Senator"

build_state_name(member)

build_state_name(String.t) :: String.t

Returns the Member’s state affiliation as a two character String.

TODO: Compare this to a map of state abbreviations/names and return

    the full state name instead

EX:

WhoWeb.MemberView.build_state_name("C001035")
>>> "ME"

parse_member_chamber(member)

parse_member_chamber(String.t) :: String.t

Parses out the Member’s chamber from the Member JSON blob.

EX:

WhoWeb.MemberView.parse_member_chamber("C001070")
>>> "Senate"

parse_member_district(member, binary)

parse_member_district(String.t, String.t) :: String.t

For all Members in the House, this function will return the Member’s district number (or “At Large” for states with only one Representative).

For Senators, the return value will be nil.

EX:

WhoWeb.MemberView.parse_member_district(_, "Senate")
>>> nil

WhoWeb.MemberView.parse_member_district("C001109", "House")
>>> "At Large"

WhoWeb.MemberView.parse_member_district("B001300", "House")
>>> "44"

parse_member_end_date(member)

parse_member_end_date(String.t) :: String.t

Returns the end date for a given Member’s current session in Congress.

EX:

WhoWeb.MemberView.parse_member_end_date("K000388")
>>> "2019-01-03"

parse_member_field(member, field)

parse_member_field(String.t, String.t) :: String.t

Extracts a given value from the top level of a Member’s JSON blob.

EX:

WhoWeb.MemberView.parse_member_field("C001070", "domain")
>>> "www.casey.senate.gov"

parse_member_name(member)

parse_member_name(String.t) :: String.t

Takes a Member’s first, middle and last name, returns them as a concatenated string.

EX:

WhoWeb.MemberView.parse_member_name("B001227")
>>> "Robert A. Brady"

WhoWeb.MemberView.parse_member_name("C001070")
>>> "Bob Casey"

parse_member_social_account(member, social_network)

parse_member_social_account(String.t, String.t) :: String.t

Returns a Member’s handle for a given social media network.

EX:

WhoWeb.MemberView.parse_member_social_account("B001227", "facebook")
>>> "RepRobertBrady"

WhoWeb.MemberView.parse_member_social_account("B001227", "twitter")
>>> "RepBrady"

WhoWeb.MemberView.parse_member_social_account("B001227", "youtube")
>>> "BradyPA01"

parse_member_start_date(member)

parse_member_start_date(String.t) :: String.t

Returns the start date from a given Member’s first session in Congress.

EX:

WhoWeb.MemberView.parse_member_start_date("K000388")
>>> "2015-06-09"

render(template, assigns \\ %{})

Renders the given template locally.

template_not_found(template, assigns)

template_not_found(Phoenix.Template.name, map) :: no_return

Callback invoked when no template is found. By default it raises but can be customized to render a particular template.