who v0.0.1 WhoWeb.MemberView

Helper methods to parse and populate %MemberPresenter{} fields.

Functions

__phoenix_recompile__?()

Returns true whenever the list of templates changes in the filesystem.

__resource__()

The resource name, as an atom, for this view

__templates__()

Returns the template root alongside all templates.

build_aggregate_party_vote_pct(member)

build_aggregate_party_vote_pct(map) :: float

Calculates the percentage of times a Member has voted with their given party by averaging this statistic from each Congressional session they’ve served in.

EX:

WhoWeb.MemberView.build_aggregate_party_vote_pct(member)
# (98.53 + 96.15) / 2 = 97.34
>>> 97.34

build_current_session_committee_list(member)

build_current_session_committee_list(map) :: nil | [String.t]

For a given member, return a list of Committee names they’re serving on in the current session.

EX:

WhoWeb.MemberView.build_current_session_committee_list(member)
>>> ["Committee on Agriculture", "Committee on Armed Services", "Committee on Small Business"]

build_member_party(member)

build_member_party(map) :: nil | String.t

Takes the “R” or “D” party flag from the Member JSON blob and returns the full party name as a string.

EX:

WhoWeb.MemberView.build_member_party(member)
>>> "Independent"

WhoWeb.MemberView.build_member_party(member)
>>> "Republican"

build_member_title(chamber)

build_member_title(String.t) :: String.t | nil

Based on the Member’s chamber, this function returns the Member’s title

EX:

WhoWeb.MemberView.build_member_title("House")
>>> "Representative"

WhoWeb.MemberView.build_member_title("Senate")
>>> "Senator"

build_social_profile_url(user_name, network)

build_social_profile_url(String.t, String.t) :: String.t

Takes in the name of a social network and the Member’s username, returns the URL to their user profile.

EX: WhoWeb.MemberView.build_social_profile_url(“SenSanders”, “twitter”) >>>

build_state_name(member)

build_state_name(map) :: nil | String.t

Returns the Member’s state affiliation as a two character String.

TODO: Compare this to a map of state abbreviations/names and return

    the full state name instead

EX:

WhoWeb.MemberView.build_state_name(member)
>>> "ME"

build_votes_list(vote_list)

build_votes_list(list) :: [map]

Extracts a list of maps from the Member’s votes JSON blob.

EX:

WhoWeb.MemberView.build_votes_list(vote_list)
>>> %{
      date: "2017-07-28",
      description: "Department of Veterans Affairs Bonus Transparency Act",
      is_bill: true,
      number: "S 114",
      position: "Yes",
      time: "12:52:00",
      title: "To authorize appropriations and to appropriate amounts for the
      Veterans Choice Program of the Department of Veterans Affairs, to improve
      hiring authorities of the Department, to authorize major medical facility
      leases, and for other purposes."
    }

parse_member_district(member, binary)

parse_member_district(map, String.t) ::
  nil |
  String.t |
  non_neg_integer

For all Members in the House, this function will return the Member’s district number (or “At Large” for states with only one Representative).

For Senators, the return value will be nil.

EX:

WhoWeb.MemberView.parse_member_district(member, "Senate")
>>> nil

WhoWeb.MemberView.parse_member_district(member, "House")
>>> "At Large"

WhoWeb.MemberView.parse_member_district(member, "House")
>>> "44"

parse_member_end_date(member)

parse_member_end_date(map) :: nil | String.t

Returns the end date for a given Member’s current session in Congress.

EX:

WhoWeb.MemberView.parse_member_end_date(member)
>>> "2019-01-03"

parse_member_field(member, key)

parse_member_field(map, String.t) :: String.t

Extracts a given value from the top level of a Member’s JSON blob.

EX:

WhoWeb.MemberView.parse_member_field(member, "domain")
>>> "www.casey.senate.gov"

WhoWeb.MemberView.parse_member_field(member, "roles", "state")
>>> "PA"

parse_member_field(member, parent_key, nested_key)

parse_member_field(map, String.t, String.t) :: String.t

parse_member_name(member)

parse_member_name(map) :: nil | String.t

Takes a Member’s first, middle and last name, returns them as a concatenated string.

EX:

WhoWeb.MemberView.parse_member_name(member)
>>> "Robert A. Brady"

WhoWeb.MemberView.parse_member_name(member)
>>> "Bob Casey"

parse_member_social_account(member, social_network)

parse_member_social_account(map, String.t) :: nil | String.t

Returns a Member’s handle for a given social media network.

EX:

WhoWeb.MemberView.parse_member_social_account(member, "facebook")
>>> "RepRobertBrady"

WhoWeb.MemberView.parse_member_social_account(member, "twitter")
>>> "RepBrady"

WhoWeb.MemberView.parse_member_social_account(member, "youtube")
>>> "BradyPA01"

parse_member_start_date(member)

parse_member_start_date(map) :: nil | String.t

Returns the start date from a given Member’s first session in Congress.

EX:

WhoWeb.MemberView.parse_member_start_date(member)
>>> "2015-06-09"

render(template, assigns \\ %{})

Renders the given template locally.

template_not_found(template, assigns)

template_not_found(Phoenix.Template.name, map) :: no_return

Callback invoked when no template is found. By default it raises but can be customized to render a particular template.